feature:S3C-4357 Split Test cases into separate classes
Description:
Split Test cases into separate classes to avoid GodClass design flaw.
ref: https://pmd.github.io/pmd-6.33.0/pmd_rules_java_design.html#godclass
